Bonus reqs
removeSong function now updates state of player.playlists as well.
Everytime you delete a song that exists in a playlist it gets delted in the playlist as well, if you delete all of the songs in the playlist the playlist itself is deleted.

When adding a new song the list will remain sorted by title.
All of the new songs are colored by their duration.